outgamelevel type 수정, 로테이션 동기화 추가, 액션 실행중 포지션동기화 하지 않도록 하는 작업중

+++ Assets/MirageTest/Editor/ActorDevModeEditorWindow.cs
@@ -58,12 +58,12 @@ public class ActorDevModeEditorWindow : OdinEditorWindow
 
     void SpawnMine(int diceId)
     {
-        _actorDevMode.SpawnMyMinion(diceId, (byte)inGameLevel, (short)outGameLevel, (byte)diceScale);
+        _actorDevMode.SpawnMyMinion(diceId, (byte)inGameLevel, (byte)outGameLevel, (byte)diceScale);
     }
     
     void SpawnEnemys(int diceId)
     {
-        _actorDevMode.SpawnEnemyMinion(diceId, (byte)inGameLevel, (short)outGameLevel, (byte)diceScale);
+        _actorDevMode.SpawnEnemyMinion(diceId, (byte)inGameLevel, (byte)outGameLevel, (byte)diceScale);
     }
     
     public class DiceElement
